// Backend-Schedra/src/modules/services/dtos/service.dto.ts
export type ServiceDto = {
  id: number;
  name: string;
  category: string;
  durationMinutes: number;
  price: number;
  description: string;
};

export type CreateServiceRequestDto = {
  name: string;
  category: string;
  durationMinutes: number;
  price: number;
  description: string;
};

export type UpdateServiceRequestDto = {
  name: string;
  category: string;
  durationMinutes: number;
  price: number;
  description: string;
};

export type ListServicesQueryDto = {
  limit?: number;
  page?: number;
  search?: string;
};